Класс TMercury20002Device
- исправлены шаблоны и типы параметров

+++ wb-mqtt-serial/mercury230_device.cpp
@@ -68,8 +68,8 @@ const TMercury230Device::TValueArray& TMercury230Device::ReadValueArray(uint32_t
         return it->second;
 
     uint8_t cmdBuf[2];
-    cmdBuf[0] = (address >> 4) & 0xff; // high nibble = array number, lower nibble = month
-    cmdBuf[1] = (address >> 12) & 0x0f; // tariff
+    cmdBuf[0] = (uint8_t)((address >> 4) & 0xff); // high nibble = array number, lower nibble = month
+    cmdBuf[1] = (uint8_t)((address >> 12) & 0x0f); // tariff
     uint8_t buf[MAX_LEN], *p = buf;
     TValueArray a;
     Talk(slave, 0x05, cmdBuf, 2, -1, buf, 16);
